pyproject.toml
[tool.poetry]
name = "FG"
version = "0.1.0"
description = "fifth-generation-megacomputer"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0a13"
nb-cli = "^0.4.0"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"

[nonebot.plugins]
plugins = ["nonebot_plugin_status", "nonebot_plugin_test"]
plugin_dirs = ["aabbcco_s_nonebot2/plugins"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
